finish tic-tac-toe game spec
