+++ cypress/component/advanced/tutorial/README.md
@@ -1,3 +1,16 @@
 React tutorial from https://reactjs.org/tutorial/tutorial.html
 
-Just following the tutorial and writing Cypress component tests
+Just following the tutorial and writing Cypress component tests.
+
+## Tic-tac-toe
+
+The example code taken from [https://codepen.io/gaearon/pen/LyyXgK](https://codepen.io/gaearon/pen/LyyXgK)
+
+<!-- prettier-ignore-start -->
+Test | Description
+--- | ---
+[square-spec.js](square-spec.js) | Testing individual square
+[tic-tac-toe-spec.js](tic-tac-toe-spec.js) | Testing the entire game
+<!-- prettier-ignore-end -->
+
+![Tic tac toe](images/tic-tac-toe.gif)
