+++ cypress/component/advanced/tutorial/square-spec.js
@@ -31,4 +31,19 @@ describe('Square', () => {
       .click()
       .should('have.text', 'X')
   })
+
+  it('looks good', () => {
+    mount(<Square />, {
+      cssFile: 'cypress/component/advanced/tutorial/tic-tac-toe.css',
+    })
+    // pause to show it
+    cy.wait(1000)
+    cy.get('.square').click()
+    cy.wait(1000)
+
+    // check if style was applied
+    cy.get('.square')
+      .should('have.css', 'background-color', 'rgb(255, 255, 255)')
+      .and('have.css', 'border', '1px solid rgb(153, 153, 153)')
+  })
 })
